# openapi/reference/zauth.v1.yaml
openapi: 3.0.0
info:
  title: Zauth
  version: '1.0'
servers:
  - url: 'http://localhost:8000'
paths:
  /users:
    get:
      summary: List users
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      operationId: get-users
      description: ''
    post:
      summary: Create user
      operationId: post-users
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
        description: ''
...
  '/users/{username}':
    get:
      summary: Show user
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      operationId: get-users-id
    parameters:
      - schema:
          type: number
        name: id
        in: path
        required: true
    put:
      summary: Update user
      operationId: put-users-id
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserChange'
  '/users/{username}/admin':
    parameters:
      - schema:
          type: number
        name: id
        in: path
        required: true
    post:
      summary: Set admin
      operationId: post-users-id-admin
      responses:
        '204':
          description: No Content
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeAdmin'
components:
  schemas:
...
    User:
      title: User
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        admin:
          type: boolean
...
    NewUser:
      title: NewUser
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
...
    UserChange:
      title: UserChange
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    ChangeAdmin:
      title: ChangeAdmin
      type: object
      properties:
        admin:
          type: boolean
  securitySchemes:
    Admin login:
      type: http
      scheme: basic
      description: ''